Use a system property to enable the unknown include to trigger changed status.
The default behavior is set so the up-to-date is based on the
hash of the file using the unknown include and not forced to
changed status if there are unknown includes. Setting the
system property org.gradle.unknown.include.strict=true will
enable the original behavior.

Make SimpliVity specific version number.